pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "drug-dataset"
description = "AI-augmented pipeline for drug discovery datasets"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Drug-Dataset Contributors"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Chemistry",
]
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "rdkit-pypi>=2022.9.1",
    "requests>=2.28.0",
    "tqdm>=4.64.0",
    "pyarrow>=10.0.0",
    "scikit-learn>=1.1.0",
    "chembl-webresource-client>=0.10.2",
    "biopython>=1.80",
    "pydantic>=1.10.0",
    "python-dotenv>=0.19.0"
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "flake8>=5.0.0",
    "black>=22.0.0",
    "isort>=5.0.0"
]

[project.urls]
Homepage = "https://github.com/KrithikVishal/Drug-Dataset"
Repository = "https://github.com/KrithikVishal/Drug-Dataset"
Issues = "https://github.com/KrithikVishal/Drug-Dataset/issues"

[project.scripts]
build-datasets = "scripts.build_datasets:main"

[tool.black]
line-length = 127
target-version = ['py38', 'py39', 'py310']

[tool.isort]
profile = "black"
line_length = 127

[tool.flake8]
max-line-length = 127
extend-ignore = ["E203", "W503"]